Extract factories and event dispatches
