+++ android/src/main/java/com/swmansion/gesturehandler/react/RNGestureHandlerModule.kt
@@ -1,35 +1,16 @@
 package com.swmansion.gesturehandler.react
 
 import android.util.Log
-import android.view.MotionEvent
 import com.facebook.react.ReactRootView
 import com.facebook.react.bridge.JSApplicationIllegalArgumentException
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactMethod
 import com.facebook.react.bridge.ReadableMap
-import com.facebook.react.bridge.WritableMap
 import com.facebook.react.module.annotations.ReactModule
-import com.facebook.react.uimanager.events.Event
 import com.facebook.soloader.SoLoader
 import com.swmansion.common.GestureHandlerStateManager
-import com.swmansion.gesturehandler.BuildConfig
 import com.swmansion.gesturehandler.NativeRNGestureHandlerModuleSpec
-import com.swmansion.gesturehandler.ReanimatedEventDispatcher
-import com.swmansion.gesturehandler.core.FlingGestureHandler
 import com.swmansion.gesturehandler.core.GestureHandler
-import com.swmansion.gesturehandler.core.HoverGestureHandler
-import com.swmansion.gesturehandler.core.LongPressGestureHandler
-import com.swmansion.gesturehandler.core.ManualGestureHandler
-import com.swmansion.gesturehandler.core.NativeViewGestureHandler
-import com.swmansion.gesturehandler.core.OnTouchEventListener
-import com.swmansion.gesturehandler.core.PanGestureHandler
-import com.swmansion.gesturehandler.core.PinchGestureHandler
-import com.swmansion.gesturehandler.core.RotationGestureHandler
-import com.swmansion.gesturehandler.core.TapGestureHandler
-import com.swmansion.gesturehandler.dispatchEvent
-
-// NativeModule.onCatalystInstanceDestroy() was deprecated in favor of NativeModule.invalidate()
-// ref: https://github.com/facebook/react-native/commit/18c8417290823e67e211bde241ae9dde27b72f17
 
 // UIManagerModule.resolveRootTagFromReactTag() was deprecated and will be removed in the next RN release
 // ref: https://github.com/facebook/react-native/commit/acbf9e18ea666b07c1224a324602a41d0a66985e
@@ -38,34 +19,10 @@ import com.swmansion.gesturehandler.dispatchEvent
 class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
   NativeRNGestureHandlerModuleSpec(reactContext), GestureHandlerStateManager {
 
-  private val eventListener = object : OnTouchEventListener {
-    override fun <T : GestureHandler<T>> onHandlerUpdate(handler: T, event: MotionEvent) {
-      this@RNGestureHandlerModule.onHandlerUpdate(handler)
-    }
-
-    override fun <T : GestureHandler<T>> onStateChange(handler: T, newState: Int, oldState: Int) {
-      this@RNGestureHandlerModule.onStateChange(handler, newState, oldState)
-    }
-
-    override fun <T : GestureHandler<T>> onTouchEvent(handler: T) {
-      this@RNGestureHandlerModule.onTouchEvent(handler)
-    }
-  }
-  private val handlerFactories = arrayOf<GestureHandler.Factory<*>>(
-    NativeViewGestureHandler.Factory(),
-    TapGestureHandler.Factory(),
-    LongPressGestureHandler.Factory(),
-    PanGestureHandler.Factory(),
-    PinchGestureHandler.Factory(),
-    RotationGestureHandler.Factory(),
-    FlingGestureHandler.Factory(),
-    ManualGestureHandler.Factory(),
-    HoverGestureHandler.Factory(),
-  )
+  private val eventDispatcher = RNGestureHandlerEventDispatcher(reactApplicationContext)
   val registry: RNGestureHandlerRegistry = RNGestureHandlerRegistry()
   private val interactionManager = RNGestureHandlerInteractionManager()
   private val roots: MutableList<RNGestureHandlerRootHelper> = ArrayList()
-  private val reanimatedEventDispatcher = ReanimatedEventDispatcher()
   override fun getName() = NAME
 
   @Suppress("UNCHECKED_CAST")
@@ -81,19 +38,16 @@ class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
       )
     }
 
-    for (handlerFactory in handlerFactories as Array<GestureHandler.Factory<T>>) {
-      if (handlerFactory.name == handlerName) {
-        val handler = handlerFactory.create(reactApplicationContext).apply {
-          tag = handlerTag
-          setOnTouchEventListener(eventListener)
-        }
-        registry.registerHandler(handler)
-        interactionManager.configureInteractions(handler, config)
-        handlerFactory.setConfig(handler, config)
-        return
-      }
+    val handlerFactory = RNGestureHandlerFactoryUtil.findFactoryForName<T>(handlerName)
+      ?: throw JSApplicationIllegalArgumentException("Invalid handler name $handlerName")
+
+    val handler = handlerFactory.create(reactApplicationContext).apply {
+      tag = handlerTag
+      setOnTouchEventListener(eventDispatcher)
     }
-    throw JSApplicationIllegalArgumentException("Invalid handler name $handlerName")
+    registry.registerHandler(handler)
+    interactionManager.configureInteractions(handler, config)
+    handlerFactory.setConfig(handler, config)
   }
 
   @ReactMethod
@@ -125,12 +79,10 @@ class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
   private fun <T : GestureHandler<T>> updateGestureHandlerHelper(handlerTag: Int, config: ReadableMap) {
     val handler = registry.getHandler(handlerTag) as T?
     if (handler != null) {
-      val factory = findFactoryForHandler(handler)
-      if (factory != null) {
-        interactionManager.dropRelationsForHandlerWithTag(handlerTag)
-        interactionManager.configureInteractions(handler, config)
-        factory.setConfig(handler, config)
-      }
+      val factory = RNGestureHandlerFactoryUtil.findFactoryForHandler(handler) ?: return
+      interactionManager.dropRelationsForHandlerWithTag(handlerTag)
+      interactionManager.configureInteractions(handler, config)
+      factory.setConfig(handler, config)
     }
   }
 
@@ -192,25 +144,6 @@ class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
 
   private external fun decorateRuntime(jsiPtr: Long)
 
-  override fun getConstants(): Map<String, Any> {
-    return mapOf(
-      "State" to mapOf(
-        "UNDETERMINED" to GestureHandler.STATE_UNDETERMINED,
-        "BEGAN" to GestureHandler.STATE_BEGAN,
-        "ACTIVE" to GestureHandler.STATE_ACTIVE,
-        "CANCELLED" to GestureHandler.STATE_CANCELLED,
-        "FAILED" to GestureHandler.STATE_FAILED,
-        "END" to GestureHandler.STATE_END
-      ),
-      "Direction" to mapOf(
-        "RIGHT" to GestureHandler.DIRECTION_RIGHT,
-        "LEFT" to GestureHandler.DIRECTION_LEFT,
-        "UP" to GestureHandler.DIRECTION_UP,
-        "DOWN" to GestureHandler.DIRECTION_DOWN
-      )
-    )
-  }
-
   override fun invalidate() {
     registry.dropAllHandlers()
     interactionManager.reset()
@@ -219,8 +152,9 @@ class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
         val sizeBefore: Int = roots.size
         val root: RNGestureHandlerRootHelper = roots[0]
         root.tearDown()
-        if (roots.size >= sizeBefore) {
-          throw IllegalStateException("Expected root helper to get unregistered while tearing down")
+
+        assert(roots.size < sizeBefore) {
+          "Expected root helper to get unregistered while tearing down"
         }
       }
     }
@@ -252,131 +186,6 @@ class RNGestureHandlerModule(reactContext: ReactApplicationContext?) :
         it.rootView is ReactRootView && it.rootView.rootViewTag == rootViewTag
       }
     }
-  }
-
-  @Suppress("UNCHECKED_CAST")
-  private fun <T : GestureHandler<T>> findFactoryForHandler(handler: GestureHandler<T>): GestureHandler.Factory<T>? =
-    handlerFactories.firstOrNull { it.type == handler.javaClass } as GestureHandler.Factory<T>?
-
-  private fun <T : GestureHandler<T>> onHandlerUpdate(handler: T) {
-    // triggers onUpdate and onChange callbacks on the JS side
-
-    if (handler.tag < 0) {
-      // root containers use negative tags, we don't need to dispatch events for them to the JS
-      return
-    }
-    if (handler.state == GestureHandler.STATE_ACTIVE) {
-      val handlerFactory = findFactoryForHandler(handler) ?: return
-
-      if (handler.actionType == GestureHandler.ACTION_TYPE_REANIMATED_WORKLET) {
-        // Reanimated worklet
-        val event = RNGestureHandlerEvent.obtain(handler, handlerFactory.createEventBuilder(handler))
-        sendEventForReanimated(event)
-      } else if (handler.actionType == GestureHandler.ACTION_TYPE_NATIVE_ANIMATED_EVENT) {
-        // Animated with useNativeDriver: true
-        val event = RNGestureHandlerEvent.obtain(
-          handler,
-          handlerFactory.createEventBuilder(handler),
-          true
-        )
-        sendEventForNativeAnimatedEvent(event)
-      } else if (handler.actionType == GestureHandler.ACTION_TYPE_JS_FUNCTION_OLD_API) {
-        // JS function, Animated.event with useNativeDriver: false using old API
-        if (BuildConfig.IS_NEW_ARCHITECTURE_ENABLED) {
-          val data = RNGestureHandlerEvent.createEventData(handlerFactory.createEventBuilder(handler))
-          sendEventForDeviceEvent(RNGestureHandlerEvent.EVENT_NAME, data)
-        } else {
-          val event = RNGestureHandlerEvent.obtain(handler, handlerFactory.createEventBuilder(handler))
-          sendEventForDirectEvent(event)
-        }
-      } else if (handler.actionType == GestureHandler.ACTION_TYPE_JS_FUNCTION_NEW_API) {
-        // JS function, Animated.event with useNativeDriver: false using new API
-        val data = RNGestureHandlerEvent.createEventData(handlerFactory.createEventBuilder(handler))
-        sendEventForDeviceEvent(RNGestureHandlerEvent.EVENT_NAME, data)
-      }
-    }
-  }
-
-  private fun <T : GestureHandler<T>> onStateChange(handler: T, newState: Int, oldState: Int) {
-    // triggers onBegin, onStart, onEnd, onFinalize callbacks on the JS side
-
-    if (handler.tag < 0) {
-      // root containers use negative tags, we don't need to dispatch events for them to the JS
-      return
-    }
-    val handlerFactory = findFactoryForHandler(handler) ?: return
-
-    if (handler.actionType == GestureHandler.ACTION_TYPE_REANIMATED_WORKLET) {
-      // Reanimated worklet
-      val event = RNGestureHandlerStateChangeEvent.obtain(handler, newState, oldState, handlerFactory.createEventBuilder(handler))
-      sendEventForReanimated(event)
-    } else if (handler.actionType == GestureHandler.ACTION_TYPE_NATIVE_ANIMATED_EVENT ||
-      handler.actionType == GestureHandler.ACTION_TYPE_JS_FUNCTION_OLD_API
-    ) {
-      // JS function or Animated.event with useNativeDriver: false with old API
-      if (BuildConfig.IS_NEW_ARCHITECTURE_ENABLED) {
-        val data = RNGestureHandlerStateChangeEvent.createEventData(handlerFactory.createEventBuilder(handler), newState, oldState)
-        sendEventForDeviceEvent(RNGestureHandlerStateChangeEvent.EVENT_NAME, data)
-      } else {
-        val event = RNGestureHandlerStateChangeEvent.obtain(handler, newState, oldState, handlerFactory.createEventBuilder(handler))
-        sendEventForDirectEvent(event)
-      }
-    } else if (handler.actionType == GestureHandler.ACTION_TYPE_JS_FUNCTION_NEW_API) {
-      // JS function or Animated.event with useNativeDriver: false with new API
-      val data = RNGestureHandlerStateChangeEvent.createEventData(handlerFactory.createEventBuilder(handler), newState, oldState)
-      sendEventForDeviceEvent(RNGestureHandlerStateChangeEvent.EVENT_NAME, data)
-    }
-  }
-
-  private fun <T : GestureHandler<T>> onTouchEvent(handler: T) {
-    // triggers onTouchesDown, onTouchesMove, onTouchesUp, onTouchesCancelled callbacks on the JS side
-
-    if (handler.tag < 0) {
-      // root containers use negative tags, we don't need to dispatch events for them to the JS
-      return
-    }
-    if (handler.state == GestureHandler.STATE_BEGAN || handler.state == GestureHandler.STATE_ACTIVE ||
-      handler.state == GestureHandler.STATE_UNDETERMINED || handler.view != null
-    ) {
-      if (handler.actionType == GestureHandler.ACTION_TYPE_REANIMATED_WORKLET) {
-        // Reanimated worklet
-        val event = RNGestureHandlerTouchEvent.obtain(handler)
-        sendEventForReanimated(event)
-      } else if (handler.actionType == GestureHandler.ACTION_TYPE_JS_FUNCTION_NEW_API) {
-        // JS function, Animated.event with useNativeDriver: false with new API
-        val data = RNGestureHandlerTouchEvent.createEventData(handler)
-        sendEventForDeviceEvent(RNGestureHandlerEvent.EVENT_NAME, data)
-      }
-    }
-  }
-
-  private fun <T : Event<T>>sendEventForReanimated(event: T) {
-    // Delivers the event to Reanimated.
-    if (BuildConfig.IS_NEW_ARCHITECTURE_ENABLED) {
-      // Send event directly to Reanimated
-      reanimatedEventDispatcher.sendEvent(event, reactApplicationContext)
-    } else {
-      // In the old architecture, Reanimated subscribes for specific direct events.
-      sendEventForDirectEvent(event)
-    }
-  }
-
-  private fun sendEventForNativeAnimatedEvent(event: RNGestureHandlerEvent) {
-    // Delivers the event to NativeAnimatedModule.
-    // TODO: send event directly to NativeAnimated[Turbo]Module
-    // ReactContext.dispatchEvent is an extension function, depending on the architecture it will
-    // dispatch event using UIManagerModule or FabricUIManager.
-    reactApplicationContext.dispatchEvent(event)
-  }
-
-  private fun <T : Event<T>>sendEventForDirectEvent(event: T) {
-    // Delivers the event to JS as a direct event. This method is called only on Paper.
-    reactApplicationContext.dispatchEvent(event)
-  }
-
-  private fun sendEventForDeviceEvent(eventName: String, data: WritableMap) {
-    // Delivers the event to JS as a device event.
-    reactApplicationContext.deviceEventEmitter.emit(eventName, data)
   }
 
   companion object {
